knative/requirements.txt
gunicorn
flask
psutil
gevent
cloudant
confluent_kafka
PyYAML
dill
pika==0.13
python-dateutil
docker
boto3
kubernetes